+++ protokt-runtime/src/jvmMain/kotlin/com/toasttab/protokt/v1/NewToOldAdapter.kt
@@ -66,10 +66,10 @@ internal class NewToOldAdapter(
     override fun readUnknown() =
         deserializer.readUnknown().let {
             when (it.value) {
-                is VarintVal -> com.toasttab.protokt.rt.UnknownField.varint(it.fieldNumber, it.value.value.value.toLong())
-                is Fixed32Val -> com.toasttab.protokt.rt.UnknownField.fixed32(it.fieldNumber, it.value.value.value.toInt())
-                is Fixed64Val -> com.toasttab.protokt.rt.UnknownField.fixed64(it.fieldNumber, it.value.value.value.toLong())
-                is LengthDelimitedVal -> com.toasttab.protokt.rt.UnknownField.lengthDelimited(it.fieldNumber, it.value.value.value)
+                is VarintVal -> com.toasttab.protokt.rt.UnknownField.varint(it.fieldNumber.toInt(), it.value.value.value.toLong())
+                is Fixed32Val -> com.toasttab.protokt.rt.UnknownField.fixed32(it.fieldNumber.toInt(), it.value.value.value.toInt())
+                is Fixed64Val -> com.toasttab.protokt.rt.UnknownField.fixed64(it.fieldNumber.toInt(), it.value.value.value.toLong())
+                is LengthDelimitedVal -> com.toasttab.protokt.rt.UnknownField.lengthDelimited(it.fieldNumber.toInt(), it.value.value.value)
                 else -> error("unsupported unknown field type")
             }
         }
