+++ protokt-runtime/src/jvmMain/kotlin/com/toasttab/protokt/v1/OldToNewAdapter.kt
@@ -66,10 +66,10 @@ internal class OldToNewAdapter(
     override fun readUnknown() =
         deserializer.readUnknown().let {
             when (it.value) {
-                is com.toasttab.protokt.rt.VarintVal -> UnknownField.varint(it.fieldNumber, it.value.value.value)
-                is com.toasttab.protokt.rt.Fixed32Val -> UnknownField.fixed32(it.fieldNumber, it.value.value.value.toUInt())
-                is com.toasttab.protokt.rt.Fixed64Val -> UnknownField.fixed64(it.fieldNumber, it.value.value.value.toULong())
-                is com.toasttab.protokt.rt.LengthDelimitedVal -> UnknownField.lengthDelimited(it.fieldNumber, it.value.value.value)
+                is com.toasttab.protokt.rt.VarintVal -> UnknownField.varint(it.fieldNumber.toUInt(), it.value.value.value)
+                is com.toasttab.protokt.rt.Fixed32Val -> UnknownField.fixed32(it.fieldNumber.toUInt(), it.value.value.value.toUInt())
+                is com.toasttab.protokt.rt.Fixed64Val -> UnknownField.fixed64(it.fieldNumber.toUInt(), it.value.value.value.toULong())
+                is com.toasttab.protokt.rt.LengthDelimitedVal -> UnknownField.lengthDelimited(it.fieldNumber.toUInt(), it.value.value.value)
                 else -> error("unsupported unknown field type")
             }
         }
